# SANGAM-PRODUCTION/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SANGAM — Supply Chain Management API",
    "version": "1.0.0",
    "description": "REST API for the SANGAM permissioned blockchain-based supply chain\nmanagement system built for the Indian Army.\n\n## Authentication\nAll protected endpoints require a `Bearer` JWT obtained from `POST /api/auth/login`.\nTokens expire after 8 hours (one operational shift). Use `POST /api/auth/refresh` to rotate.\n\n## Permissions\nEvery endpoint lists the required permission (format: `resource:action`).\nPermissions are granted by role (Day 13 RBAC), temporary delegation (Day 15),\nor single-use emergency override (Day 15).\n\n## Command Scope\nData is filtered to the requesting user's command scope — their unit and all\nsubordinate units. A battalion CO automatically sees all companies and platoons.\nLateral access (sibling units) is denied.\n\n## Offline-First\nAll services degrade gracefully when the database is unavailable.\nNotifications and audit events are buffered in memory and flushed on reconnect.",
    "contact": {
      "name": "SANGAM Development Team",
      "url": "https://github.com/sangam-scm"
    },
    "license": {
      "name": "Proprietary — Indian Army Internal Use"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local / Docker development"
    },
    {
      "url": "https://sangam.army.mil",
      "description": "Production (classified network)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/auth/login (Day 14). Expires after 8h."
      }
    },
    "schemas": {
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "A descriptive error message"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "jco_ram"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Str0ngPass!"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "accessToken": {
            "type": "string",
            "description": "JWT (8h lifetime)"
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque 128-char hex (30d lifetime)"
          },
          "refreshExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "username": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              },
              "role": {
                "$ref": "#/components/schemas/RoleName"
              },
              "unitId": {
                "type": "integer"
              },
              "unitCode": {
                "type": "string"
              }
            }
          }
        }
      },
      "RoleName": {
        "type": "string",
        "enum": [
          "SOLDIER",
          "NCO",
          "JCO",
          "LOGISTICS_OFFICER",
          "OFFICER",
          "SENIOR_OFFICER",
          "COMMANDER",
          "AUDITOR",
          "SYSTEM_ADMIN"
        ],
        "description": "Indian Army rank-mapped role (Day 13)"
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions or outside command scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticate with username/password. Returns JWT access token (8h) and refresh token (30d). Rate-limited: 10 attempts per IP per 5 minutes.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "423": {
            "description": "Account locked (5 consecutive failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rotate refresh token",
        "description": "Exchange a refresh token for a new access+refresh pair (single-use rotation). Presenting a revoked token triggers a security event and revokes all sessions.",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout (single session)",
        "description": "Revoke a specific refresh token.",
        "operationId": "logout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout (all sessions)",
        "description": "Revoke all refresh tokens for the authenticated user.",
        "operationId": "logoutAll",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Change password",
        "description": "Change password. Requires current password. On success, all sessions are revoked (force re-login everywhere).",
        "operationId": "changePassword",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "oldPassword",
                  "newPassword"
                ],
                "properties": {
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "Min 8 chars, 1 uppercase, 1 lowercase, 1 digit"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/unlock/{userId}": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Admin unlock account",
        "description": "Clear account lockout for a user. Requires users:write permission.",
        "operationId": "unlockAccount",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account unlocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Current user identity",
        "description": "Return the identity and role of the authenticated user.",
        "operationId": "getMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "username": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "$ref": "#/components/schemas/RoleName"
                        },
                        "rankLevel": {
                          "type": "integer"
                        },
                        "unitId": {
                          "type": "integer"
                        },
                        "unitCode": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "security": []
}